Members portfolio dashboard

+++ app/Http/Controllers/Members/BasketController.php
@@ -3,9 +3,6 @@
 namespace App\Http\Controllers\Members;
 
 use App\Models\AssetsManagement\Basket;
-use App\Models\AssetsManagement\StockType;
-use App\Models\Company;
-use App\Models\FiscalYear;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
@@ -69,9 +66,63 @@ class BasketController extends Controller
 
     public function fetch(Request $request)
     {
-        $dt = Datatables::of(Basket::with('stockBuy.sell')->where('user_id',Auth::id()))->make(true);
+      $lastPrices = "select `price`.`company_id` as `cid`, `price`.`date` as `close_date`
+                    , `price`.`close` as `close_price` from `todays_price` as `price`
+                    INNER JOIN `last_traded_price` as `ltp`
+                    on `price`.`company_id` = `ltp`.`company_id`
+                    and `price`.`date` = `ltp`.`date`";
 
-        return $dt;
+      $stockWithLastPrice = \DB::table('am_stock_basket as basket')
+      ->leftjoin('am_stocks_buy as stock', 'basket.id', '=', 'stock.basket_id')
+      ->leftJoin('am_stocks_sell as sell', 'sell.buy_id', '=','stock.id')
+      ->leftjoin('company', 'stock.company_id', '=','company.id')
+      ->leftJoin(\DB::raw('('.$lastPrices.') as latest_price'),'stock.company_id','=','latest_price.cid')
+      ->selectRaw('
+          (ifnull(CAST(sum(sell.quantity) AS SIGNED), 0)) as sell_quantity,
+          (stock.quantity - ifnull(CAST(sum(sell.quantity) AS SIGNED),0)) as remaining_quantity,
+          (stock.commission / stock.quantity) as commission_per_quantity,
+          stock.commission, stock.buy_rate, stock.quantity as buy_quantity,
+          latest_price.close_date, latest_price.close_price, stock.company_id, stock.id,
+          basket.id as basket_id, basket.name as basket_name, company.name as company_name
+          '
+      )
+      ->groupBy('stock.id')
+      ->where('basket.user_id', auth()->id());
+
+      $stockWithComputedValues = \DB::table(
+          \DB::raw('('.$stockWithLastPrice->toSql().') as stock')
+      )
+      ->addBinding($stockWithLastPrice->getBindings())
+      ->selectRaw('
+          stock.*,
+          (@investment:= (
+              (remaining_quantity * buy_rate)  + (commission_per_quantity * remaining_quantity)
+          )) as investment,
+          @market_value:= (remaining_quantity * close_price) as market_value,
+          @market_value - @investment as profit_loss
+      ');
+
+      $stockGroupedByCompany = \DB::table(
+          \DB::raw('('.$stockWithComputedValues->toSql().') as sc')
+      )
+      ->addBinding($stockWithComputedValues->getBindings())
+      ->selectRaw('
+          basket_id as id, basket_name as name, company_name, company_id,
+          sum(investment) as investment, sum(market_value) as market_value, sum(profit_loss) as profit_loss
+      ')
+      ->groupBy('basket_id')
+      ->groupBy('company_id');
+
+      $stockGroupedByBasket = \DB::table(
+          \DB::raw('('.$stockGroupedByCompany->toSql().') as basket')
+      )
+      ->addBinding($stockGroupedByCompany->getBindings())
+      ->selectRaw('
+          id, name, sum(investment) as investment, sum(market_value) as market_value, sum(profit_loss) as profit_loss
+      ')
+      ->groupBy('id');
+
+      return $dt = Datatables::of($stockGroupedByBasket)->make(true);
     }
 
     public function destroy(Request $request, $id)
@@ -81,12 +132,18 @@ class BasketController extends Controller
         if(is_null($basket)){
             $message = 'Basket not found.';
         }else{
+          $stock_count = $basket->stock()->count();
+
+          if($stock_count == 0) {
             if($basket->delete()){
-                $error = false;
-                $message = 'Basket has been successfully deleted.';
+              $error = false;
+              $message = 'Basket has been successfully deleted.';
             }else{
-                $message = 'Basket could not be deleted.';
+              $message = 'Basket could not be deleted.';
             }
+          }else {
+            $message = 'Basket cannot be deleted. Delete stocks first.';
+          }
         }
 
         if($request->isXmlHttpRequest()){
